chore: move custom steps into separate component
