+++ src/features/multivariate_layer/components/MultivariateAnalysisForm/MultivariateAnalysisForm.module.css
@@ -50,22 +50,6 @@
   justify-content: center;
 }
 
-.customStepsRow {
-  width: 100%;
-  display: flex;
-  flex-flow: row wrap;
-  gap: var(--unit);
-  align-items: center;
-}
-
-.customStepsRowName {
-  width: 45px;
-}
-
-.customStepInput {
-  width: 100px;
-}
-
 @media screen and (max-width: 960px) {
   .modalDialog {
     height: 85vh;
